Unify ssim calculation
